+++ src/onegov.swissvotes/onegov/swissvotes/layouts/default.py
@@ -16,6 +16,8 @@ class DefaultLayout(ChameleonLayout):
 
     def __init__(self, model, request):
         super().__init__(model, request)
+        self.request.include('frameworks')
+        self.request.include('chosen')
         self.request.include('common')
 
         self.pages = TranslatablePageCollection(self.request.session)